Add Vite generated frontend
The code in the webapp folder was mostly generated with

`npm create vite@latest`

and selecting react and typescript as our choices.

I also followed the example from the gradle node plugin:

https://github.com/node-gradle/gradle-node-plugin/blob/master/examples/vertx-react/webapp/build.gradle

which has a neat hack for creating a jar file with resources generated by the build.

I needed to update the output location for the Vite build to build/webapp/webroot
which unlike the gradle node plugin example was in vite.config.ts, but after
that it seemed to work as expected.